Strong compiler checks -Wextra and -Wconversion.  Fix warnings.

+++ src/syncAgent/types.h
@@ -7,17 +7,41 @@
  */
 
 /*
- * Delta time for adjusting SyncPoint.  2-bytes, 16-bits, 128k
- * Content of SyncMessage to adjust sync.
+ * Type of field of SyncMessage to adjust SyncPoint.
+ *
+ * Design considerations:
+ * This is transmitted over-the-air; to keep packets short, it should be small.
+ * But the platform RTC clock (OSTime, typically 24 or 32 bits) limits the max size;
+ * DeltaSync shouldn't be any larger than the platform RTC allows for timeouts.
+ *
+ * The size can be no smaller than determined by a combination of:
+ * - resolution of the platform RTC clock (OSClock)
+ * - SlotDuration
+ * - DutyCycleInverse
+ * As illustrated by the following derivation:
+ * MaxSyncPeriod == MaxDeltaSync / 2    (we must schedule two SyncPeriods forward)
+ * MaxSlotCount = MaxSyncPeriod / SlotDuration
+ * MaxDutyCycle == 3 awake slots / MaxSlotCount
+ * MaxDutyCycleInverse = 1 / MaxDutyCycle
+ *
+ * Example design choices:
+ * for OSClock resolution==1/32khz and SlotDuration==300 ticks:
+ * - 2-bytes, 16-bits, MaxDeltaSync is 64k:  max DutyCyleInverse is 30
+ * - 3-bytes, 24-bits, MaxDeltaSync is 16M: max DutyCycleInverse is 10k
+ *
+ * Note that the code might not be correct in coercing to this type.
  */
-// TODO assertions on overflows DeltaSync
-// Currently we are transmitting 3 bytes of DeltaSync
-typedef uint16_t DeltaSync;
+// TODO correct code to catch loss of data in converting to DeltaSync
+
+// Currently we are transmitting 3 bytes of DeltaSync,
+// i.e. the primitive type uint32_t is the same as for DeltaTime
+typedef uint32_t DeltaSync;
 
 
 /*
  * Used in scheduling to count slots and periods.
  */
+// TODO since we aren't enforcing this, we might as well use native int
 typedef uint16_t ScheduleCount;
